Changed how the navigation works on the page by using new pages for each clickable link.
