--- Dev/FrontEnd/front-end/app/cart.tsx
@@ -1,11 +1,0 @@
-function Cart()
-{
-    return(
-        <div>
-            <input type="text" placeholder="Email"></input>
-            <input type="password" placeholder="Password"></input>
-        </div>
-    )
-};
-
-export default Cart;
