+++ Dev/FrontEnd/front-end/app/cart/Cart.module.css
@@ -8,9 +8,3 @@
     justify-content: center;
     align-items: center;
 }
-
-.cart input
-{
-    width: 100%;
-    padding: 1rem;
-}
