+++ Dev/FrontEnd/front-end/app/page.tsx
@@ -1,9 +1,12 @@
 "use client";
+import { BrowserRouter, Routes, Route, NavLink } from 'react-router-dom';
 import { useState, useEffect } from "react";
 import Navbar from "@/components/Navbar/Navbar";
 import styles from "./page.module.css";
 import Login from "@/components/Login/Login";
-import AccountMenu from "@/components/AccountMenu/AccountMenu";
+import AccountMenu from "@/app/AccountMenu/page";
+
+
 export default function Home() {
   const [isLogin, setIsLogin] = useState(false);
   const [showLogin, setShowLogin] = useState(false);
@@ -11,10 +14,13 @@ export default function Home() {
 
   return (
     <main className={styles.main}>
-      <Navbar showLogin = {showLogin} setShowLogin={setShowLogin} isLogin={isLogin} setIsLogin={setIsLogin}
-       isAccountClick={isAccountClick} setIsAccountClick={setIsAccountClick}/>
+
+      <Navbar showLogin={showLogin} setShowLogin={setShowLogin} isLogin={isLogin} setIsLogin={setIsLogin}
+        isAccountClick={isAccountClick} setIsAccountClick={setIsAccountClick} />
       {showLogin && <Login />}
-      {isAccountClick && <AccountMenu />}
+        {isAccountClick && <AccountMenu />}
+
     </main>
   );
+
 }
