+++ Dev/FrontEnd/front-end/components/Navbar/Navbar.tsx
@@ -1,7 +1,8 @@
-import AccountPageStyles from "../AccountMenu/AccountMenu";
+import AccountPageStyles from "../../app/AccountMenu/page";
 import NavbarStyles from "./Navbar.module.css";
 import Login from "@/components/Login/Login";
-import AccountMenu from "@/components/AccountMenu/AccountMenu";
+import AccountMenu from "@/app/AccountMenu/page";
+import Link from 'next/link';
 function Navbar({showLogin, setShowLogin, isLogin, setIsLogin, isAccountClick, setIsAccountClick} : {showLogin: boolean, setShowLogin: any, isLogin: boolean, setIsLogin: any, isAccountClick: boolean, setIsAccountClick: any})
 {
     const handleLogin = () => {
@@ -10,6 +11,7 @@ function Navbar({showLogin, setShowLogin, isLogin, setIsLogin, isAccountClick, s
             console.log("Logout");
             setShowLogin(false);
             setIsLogin(false);
+            setIsAccountClick(false);
         }
         else
         {
@@ -30,19 +32,30 @@ function Navbar({showLogin, setShowLogin, isLogin, setIsLogin, isAccountClick, s
         }
     }
 
-    return(
-        <div className={NavbarStyles.navbar}>
+
+    return (
+
+<div className={NavbarStyles.navbar}>
             <h1>NotTheSilkRoad</h1>
             
             <div className={NavbarStyles.search}>
+
                 <input type="text" placeholder="Search"></input>
                 <button>Search</button>
-            </div>
             
-            <button onClick={handleAccountButton}>{isLogin ? "Account" : ""}</button>
-            <button onClick={handleLogin}>{isLogin ? "Logout" : "Login"}</button>
+            </div>
+
+          {isLogin ? (
+            <nav>
+              <Link href="/AccountMenu">Account</Link>
+              <button onClick={handleLogin}>Logout</button>
+            </nav>
+          ) : (
+            <button onClick={handleLogin}>Login</button>
+          )}
+
         </div>
-    )
-};
+      );
+    }
 
 export default Navbar;
